requirements.txt
# Traffic Light Detection - Phase 3 Requirements
# Core ML libraries
torch>=2.0.0
torchvision>=0.15.0
ultralytics>=8.0.0

# Image processing and augmentation
opencv-python>=4.5.0
Pillow>=9.0.0
albumentations>=1.3.0
numpy>=1.21.0

# Data visualization and analysis
matplotlib>=3.5.0
seaborn>=0.11.0
pandas>=1.5.0

# Utilities
tqdm>=4.64.0
pyyaml>=6.0
scikit-learn>=1.1.0